Fix native integration to pass cookies/credentials for "Extensions" info graphql request
