+++ client/browser/src/shared/backend/requestGraphQl.ts
@@ -102,7 +102,12 @@ export function createGraphQLHelpers(sourcegraphURL: string, isExtension: boolea
      * After that, the same instance should be used by all consumers.
      */
     const getBrowserGraphQLClient = once(() =>
-        getGraphQLClient({ headers: getHeaders(), baseUrl: sourcegraphURL, isAuthenticated: false })
+        getGraphQLClient({
+            headers: getHeaders(),
+            baseUrl: sourcegraphURL,
+            isAuthenticated: false,
+            credentials: 'include',
+        })
     )
 
     return { getBrowserGraphQLClient, requestGraphQL }
